requirements.txt
python-telegram-bot
simplesqlitewrap
requests
toml
pyyaml
beautifulsoup4
